requirements.txt
asgiref==3.2.7
Django==3.0.7
django-braces==1.14.0
django-cors-headers==3.3.0
django-extensions==2.2.9
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
